removed seed function and seeddb route
