+++ helpers.js
@@ -18,16 +18,16 @@ client.on('error', error => errorHandler(error));
 function errorHandler(error, message, res) {
   console.error(error);
   if (message) {
-  
     if (res) {
       res.send(message);
     }
+    console.log(message);
   }
 }
 
 // Search function
 function search(request, response) {
-  
+
   // check database to see if data is in DB
   let city = request.body.query;
   console.log(city);
@@ -43,7 +43,7 @@ function search(request, response) {
       if (locationResult.rowCount > 0) {
         //transfer control over to map function -> which will do the rendering
         location = locationResult.rows[0];
-  
+
 
         let sql = `SELECT * FROM breweries WHERE location_id = $1;`;
         let values = [location.id];
@@ -53,16 +53,16 @@ function search(request, response) {
 
             if (breweryResults.rowCount > 0) {
 
-             
+
               getBreweriesWeWantToRender(breweryResults.rows, location, response);
 
             } else { // get breweries from API
 
               let url = `https://api.brewerydb.com/v2/search/geo/point?lat=${location.lat}&lng=${location.long}&key=${process.env.BREWERYDB_API_KEY}&radius=20`;
-            
+
               superagent.get(url)
                 .then(breweryResults => {
-               
+
                   if (!breweryResults.body.data) {
                     return errorHandler({ status: 404, line: 64 }, 'No data from brewerydb', response);
                   }
@@ -71,7 +71,7 @@ function search(request, response) {
 
                     let sql = `INSERT INTO breweries(id, brewery, website, image, lat, long, time_stamp) VALUES($1, $2, $3, $4, $5, $6, $7) ON CONFLICT DO NOTHING;`;
                     let values = Object.values(brewery);
-                
+
 
                     client.query(sql, values)
                       .catch(error => errorHandler(error));
@@ -84,7 +84,7 @@ function search(request, response) {
                         // this is where the brewery gets returned for the map method
                         getBreweriesWeWantToRender(breweryQueryResult.rows, location, response);
                       })
-                      .catch(error => {errorHandler(error)})
+                      .catch(error => { errorHandler(error) })
                   });
 
                 })
@@ -130,7 +130,7 @@ function search(request, response) {
                         superagent.get(url)
                           .then(breweryResults => {
                             if (!breweryResults.body.data) {
-                              errorHandler({status: 404, line: 133}, 'No data from brewerydb', response);
+                              errorHandler({ status: 404, line: 133 }, 'No data from brewerydb', response);
                             }
                             breweries = breweryResults.body.data.map(breweryData => {
                               let brewery = new constructor.Brewery(breweryData);
@@ -191,7 +191,7 @@ function getBreweriesWeWantToRender(breweries, location, response) {
   // console.log(breweries)
   console.log(location)
   // return getBreweries(location, breweries)
-  response.render('search', {location:location, breweries: breweries});
+  response.render('search', { location: location, breweries: breweries });
 }
 //
 
@@ -212,7 +212,7 @@ function breweries(request, response) {
       console.log(`\n\n\nNO DATA FROM DB, BIG PROBLEMS\n\n\nGetting beers for this brewery from BreweryDB`);
       return beersByBreweryFromApi(brewery_id);
     } else {
-      response.render('pages/breweryDetails', {breweryBeers: breweryResult.rows});
+      response.render('pages/breweryDetails', { breweryBeers: breweryResult.rows });
     }
 
   }).then(beersFromApi => {
@@ -225,7 +225,7 @@ function breweries(request, response) {
         ibu: 0,
         time_stamp: Date.now(),
         style: { id: 1 },
-        breweries: [{id: brewery_id }]
+        breweries: [{ id: brewery_id }]
       }];
     }
 
@@ -247,7 +247,7 @@ function breweries(request, response) {
 
 function beersByBreweryFromApi(brewery_id) {
   console.log('getting beer list from api for brewery_id ', brewery_id);
-  let url=`https://api.brewerydb.com/v2/brewery/${brewery_id}/beers?withBreweries=Y&key=${process.env.BREWERYDB_API_KEY}`;
+  let url = `https://api.brewerydb.com/v2/brewery/${brewery_id}/beers?withBreweries=Y&key=${process.env.BREWERYDB_API_KEY}`;
   console.log(url);
   return superagent.get(url);
 }
@@ -276,14 +276,14 @@ function beers(request, response) {
       else reviewsArray = reviewsResult.rows;
       // console.log(`${beer}\n\n\n ${reviewsArray}`);
       console.log(reviewsResult.rows[0], request.params.beer_id);
-      response.render('pages/beerdetails', {beer: beer, reviews: reviewsArray});
+      response.render('pages/beerdetails', { beer: beer, reviews: reviewsArray });
     })
     .catch(error => errorHandler(error));
 }
 
 //update a beer entry w/ a comment
 
-function review(request, response){
+function review(request, response) {
   // let {beer_id, note, rating, time_stamp, gf} = request.body;
   console.log(request.body);
 
@@ -304,7 +304,7 @@ function review(request, response){
 
 //delete a comment
 
-function removeReview(request, response){
+function removeReview(request, response) {
   let sql = `DELETE FROM reviews WHERE id=$1`;
   let values = [request.params.review_id];
   console.log(request.body);
@@ -312,96 +312,19 @@ function removeReview(request, response){
   client.query(sql, values)
     .then(response.redirect(`/beers/${request.params.beer_id}`))
     .catch(errorHandler);
-
 }
 
-//render shelf
-
-function shelf(request, response){
-  let sql = //join function
-
-client.query(sql [request.params.beer_id]).then(shelfResult =>{
-  if(shelfResult.rows.length > 0){
-    return response.render(`./shelf`, {shelf: shelfResult.rows});
-  }
-}).catch(error => errorHandler(error));
-}
-
-//database seeding
-
-function seed(req, res) {
-  let location;
-  let sql = `SELECT * FROM locations WHERE search_query=$1;`;
-  let values = ['seattle'];
-  client
-    .query(sql, values)
-    .then(result => {
-      if (!result.rowCount) throw 'All broken, stop now';
-      location = result.rows[0];
-
-      const brewerySeed = require('./data/breweries-seattle.json').data;
-      const breweryArray = brewerySeed
-        .filter(brewery => brewery.openToPublic === 'Y')
-        .map(element => {
-
-          let brewery = new constructor.Brewery(element);
-          brewery.location_id = location.id;
-
-          let sql = `INSERT INTO breweries(id, brewery, website, image, lat, long, time_stamp, location_id) VALUES($1,$2,$3,$4,$5,$6,$7,$8) ON CONFLICT DO NOTHING;`;
-          let values = Object.values(brewery);
-
-          client.query(sql, values)
-            .catch(errorHandler);
-          //.log('🍺 Insert Complete');
-          return brewery;
-        });
-
-      const styles = require('./data/styles.json').data;
-      const styleArray = styles.map(style => {
-
-        let thisStyle = new constructor.Style(style);
-
-        let sql = `INSERT INTO styles(id, name, description, abvmin, abvmax, ibumin, ibumax, time_stamp) VALUES ($1,$2,$3,$4,$5,$6,$7,$8) ON CONFLICT DO NOTHING;`;
-        let values = Object.values(thisStyle);
-
-        client.query(sql, values)
-          .catch(errorHandler);
-        //console.log('Style Insert Complete');
-        return thisStyle;
-      });
-
-      let beerSeed, beerArray;
-      breweryArray.forEach(brewery => {
-
-        beerSeed = require(`./data/${brewery.id}.json`).data;
-        beerArray = beerSeed.map(element => {
-
-          let beer = new constructor.Beer(element);
-
-          let sql = `INSERT INTO beers(name, beer_id, abv, ibu, time_stamp, style_id, brewery_id) VALUES($1,$2,$3,$4,$5,$6,$7) ON CONFLICT DO NOTHING;`;
-          let values = Object.values(beer);
-          
-          client.query(sql, values)
-            .catch(errorHandler);
-          return beer;
-        });
-      });
-      res.render('pages/datadisplay', { breweries: breweryArray, styles: styleArray, beers: beerArray });
-    })
-    .catch(errorHandler);
-
-}
 
 function getLocation(request, response) {
   let search_query = request.query.search_query;
-  search_query='seattle';
+  search_query = 'seattle';
   let sql = `SELECT * FROM locations WHERE search_query=$1;`;
   let values = [search_query];
 
   client
     .query(sql, values)
     .then(result => {
-      if (result.rowCount === 0){
+      if (result.rowCount === 0) {
         let url = `https://maps.googleapis.com/maps/api/geocode/json?key=${process.env.GOOGLE_API_KEY}&address=${city}`;
         console.log(url)
         // return next async call to promise chain
@@ -422,7 +345,8 @@ function getLocation(request, response) {
         let sql = `INSERT INTO locations(search_query, lat, long) VALUES($1, $2, $3) RETURNING *;`;
         let values = Object.values(location);
         return client.query(sql, values);
-      }})
+      }
+    })
     .then(locationResult => {
       location = locationResult.rows[0];
       response.send(location);
@@ -430,7 +354,8 @@ function getLocation(request, response) {
     .catch(errorHandler);
 }
 
-function getBreweries (request, response) {
+// gets a list of breweries from a 
+function getBreweries(request, response) {
   // let location = request.query.location;
   // console.log('+_+_+_+_+')
   console.log(location)
@@ -440,17 +365,10 @@ function getBreweries (request, response) {
 
   let sql = `SELECT * FROM breweries WHERE location_id=$1;`;
   let values = [location.id];
-  client.query(sql, values)
-    .then(breweriesResult => {
-      // console.log(breweriesResult)
-      if (breweriesResult.rowCount === 0) {
-        // TODO: get data from api
-      }
-      else {
-        response.send({ breweries: breweriesResult.rows, location:location});
-      }
-    })
+  client.query(sql, values).then(breweriesResult => {
+    response.send({ breweries: breweriesResult.rows, location: location });
+  })
     .catch(error => errorHandler(error));
 }
 
-module.exports = {search, errorHandler, breweries, beers, seed, review, removeReview, shelf, getLocation, getBreweries};
+module.exports = { search, errorHandler, breweries, beers, seed, review, removeReview, shelf, getLocation, getBreweries };
