+++ server.js
@@ -34,18 +34,14 @@ app.listen(PORT, () => console.log(`\n\n**#############\n\nBeerly here on PORT $
 
 //server routes
 app.get('/', mainPage);
-app.post('/search', helpers.search);
-app.get('/breweries/:brewery_id', helpers.breweries);
-app.get('/beers/:beer_id', helpers.beers);
-app.get('/seeddb', helpers.seed);
-app.post('/reviews/:beer_id', helpers.review);
-app.delete('/reviews/:review_id/:beer_id', helpers.removeReview);
-app.get('/shelf/:beer_id', helpers.shelf);
-app.get('/location', helpers.getLocation);
-app.get('/breweries', helpers.getBreweries);
+app.post('/search', helpers.search); // search route, takes a location string in the request.body.and displays the map page
+app.get('/breweries/:brewery_id', helpers.breweries); // displays a specific brewery's details and beers available.
+app.get('/beers/:beer_id', helpers.beers); // displays a specific beer's details
+app.post('/reviews/:beer_id', helpers.review); // adds a review for a given beer_id
+app.delete('/reviews/:review_id/:beer_id', helpers.removeReview); // deletes the specified review and redirects to the given beer_id page
+app.get('/location', helpers.getLocation); // get a location from Google's geocode API and sends it back
+app.get('/breweries', helpers.getBreweries); // get a list of breweries for a location, needs location string, send as request.query.search_query
 app.get('/about', (request, response) => response.render('about'))
-// app.get('/beers', helpers.getBeers);
-// app.get('/styles', helpers.getStyles);
 
 //generic route for all incorrect access
 app.use('*', (req, res) => helpers.errorHandler({status: 404, line: 45, server: true}, 'You have reached a page that does not exist.', res));
